Update early access SDK 3.0.9-build20210514
Feature:
- add Arduboy zip library
API:
- update PubSubClient lib
- update IR
- update jtag enable functions
- update wifi security option
- update SDK libs and header files for latest version
Misc:
- rename folder name to short the length of path

+++ Arduino_package/hardware/system/component/os/freertos/freertos_v10.2.0/Source/portable/MemMang/heap_5.c
@@ -68,7 +68,6 @@
  *
  */
 #include <stdlib.h>
-#include <string.h>
 
 /* Defining MPU_WRAPPERS_INCLUDED_FROM_API_FILE prevents task.h from redefining
 all the API functions to use the MPU wrappers.  That should only be done when
@@ -507,8 +506,8 @@ void* pvPortReAlloc( void *pv,  size_t xWantedSize )
 			/* This casting is to keep the compiler from issuing warnings. */
 			pxLink = ( void * ) puc;
 
-			size_t oldSize =  (pxLink->xBlockSize & ~xBlockAllocatedBit) - xHeapStructSize;
-			size_t copySize = ( oldSize < xWantedSize ) ? oldSize : xWantedSize;
+			int oldSize =  (pxLink->xBlockSize & ~xBlockAllocatedBit) - xHeapStructSize;
+			int copySize = ( oldSize < xWantedSize ) ? oldSize : xWantedSize;
 			memcpy( newArea, pv, copySize );
 
 			vTaskSuspendAll();
